requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
soundfile==0.12.1
numpy==1.26.4
requests==2.32.3
chatterbox-tts==0.1.2